# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Image:
    properties:
      createdAt:
        type: string
      filename:
        type: string
      id:
        type: integer
    type: object
...
  v1.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  v1.UploadFileResponse:
    properties:
      id:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API server
  title: FinlandShop API
  version: "0.5"
paths:
...
  /images/:
    get:
      consumes:
      - application/json
      description: get all images
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Image'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      - AdminAuth: []
      summary: Get all images
      tags:
      - images-actions
    post:
      consumes:
      - application/json
      description: upload image
      parameters:
      - description: photo to upload
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UploadFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      - AdminAuth: []
      summary: Upload image
      tags:
      - images-actions
  /images/{id}:
    delete:
      consumes:
      - application/json
      description: delete image by id
      parameters:
      - description: image id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - UsersAuth: []
      - AdminAuth: []
      summary: Delete image
      tags:
      - images-actions
...
securityDefinitions:
  AdminAuth:
    in: context
    name: Admin authorization
    type: apiKey
  UsersAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"